requirements.txt
AdvancedHTTPServer==2.2.0
aiohttp==3.7.4
aiohttp-socks==0.5.3
aiorpcX==0.18.5
alembic==1.4.3.dev0
aniso8601==8.1.1
apparmor==2.13.6
appdirs==1.4.4
argcomplete==1.8.1
argon2-cffi==18.3.0
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
backdoor-factory==0.0.0
basemap==1.2.1
bcrypt==3.1.7
beautifulsoup4==4.9.3
binwalk==2.2.1
bitstring==3.1.7
bleach==3.2.1
blinker==1.4
boltons==19.1.0
cairocffi==1.2.0
capstone==4.0.2
Cerberus==1.3.2
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
chirp==0.3.0.dev0
click==7.1.2
click-plugins==1.1.1
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
cryptography==3.3.2
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.21
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
dhcpig==0.0.0
distlib==0.3.1
distro==1.5.0
distro-info==1.0
dnslib==0.9.14
dnspython==2.0.0
ecdsa==0.16.1
Electrum==4.0.9
email-validator==1.1.1
entrypoints==0.3
fierce==1.4.0
filelock==3.0.12
Flask==1.1.2
Flask-HTTPAuth==3.2.4
future==0.18.2
geoip2==2.9.0
geojson==2.5.0
gevent==20.9.0
goofile==1.6
google==2.0.3
gpg===1.14.0-unknown
graphene==2.1.7
graphene-sqlalchemy==2.1.2
graphql-core==2.2.1
graphql-relay==2.0.0
greenlet==0.4.17
gyp==0.1
hashID==3.1.4
helpdev==0.7.1
html5lib==1.1
httplib2==0.18.1
hyperlink==19.0.0
icalendar==4.0.3
idna==2.10
imageio==2.9.0
impacket==0.9.22
importlib-metadata==1.6.0
incremental==17.5.0
iniconfig==0.0.0
invoke==1.4.1
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
ldap3==2.8.1
ldapdomaindump==0.9.3
LibAppArmor==2.13.6
logutils==0.3.3
lxml==4.6.3
lz4==3.1.3+dfsg
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==1.1.1
mate-menu==20.4.1
mate-tweak==21.4.2
matplotlib==3.3.4
maxminddb==2.0.3
mechanize==0.4.5
mercurial==5.6.1
mistune==0.8.4
mock==4.0.3
more-itertools==4.2.0
msgpack==1.0.0
multidict==5.1.0
mutagen==1.45.1
mysqlclient==1.4.4
nbconvert==5.6.1
nbformat==5.1.2
netaddr==0.7.19
netifaces==0.10.9
networkx==2.5.1
nose==1.3.7
notebook==6.2.0
numpy==1.19.5
nyx==2.1.0
olefile==0.46
onboard==1.4.1
onioncircuits==0.7
onionshare==2.2
opencv-python==4.5.1.48
packaging==20.9
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.1
pbr==5.5.0
pcapy==0.11.4
pefile==2019.4.18
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.0
pluginbase==1.0.0
ply==3.11
prometheus-client==0.9.0
promise==2.2
prompt-toolkit==3.0.14
protobuf==3.12.4
psutil==5.8.0
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.16.2
pycountry==20.7.3
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.9.7
pycups==2.0.1
pycurl==7.43.0.6
pyenchant==3.2.0
Pygments==2.7.1
PyGObject==3.38.0
PyHamcrest==1.9.0
pyinotify==0.9.6
pymongo==3.11.0
pymssql==2.1.4
PyNaCl==1.4.0
pynvim==0.4.2
pyodbc==4.0.30
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
pyotp==2.3.0
pyparsing==2.4.7
pyproj==3.0.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyqtgraph==0.11.1
pyrsistent==0.15.5
pyserial==3.5b0
pyshp==2.1.3
PySimpleSOAP==1.16.2
pysmbc==1.0.23
pysmi==0.3.2
pysnmp==4.4.12
PySocks==1.7.1
pytest==6.0.2
python-apt==2.1.7
python-dateutil==2.8.1
python-debian==0.1.39
python-debianbts==3.1.0
python-editor==1.0.3
python-libnmap==0.7.2
python-magic==0.4.20
python-magic-ahupp==0.4.13
python-memcached==1.58
python-pam==1.8.4
python-xapp==2.0.2
python-xlib==0.29
pytz==2021.1
PyWavelets==1.1.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.3.1
pyzmq==20.0.0
QDarkStyle==2.8.1
qrcode==6.1
QtPy==1.9.0
redis==3.5.3
reportbug==7.10.3
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
rule-engine==1.1.0
Rx==3.1.1
scapy==2.4.4
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
selenium==4.0.0a1
Send2Trash==1.6.0b1
service-identity==18.1.0
setproctitle==1.2.1
shodan==1.25.0
simplejson==3.17.2
six==1.15.0
sk-video==1.1.10
smoke-zephyr==2.0.1
soupsieve==2.2.1
SQLAlchemy==1.3.22
stem==1.8.0
tabulate==0.8.7
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.1
torbrowser-launcher==0.3.3
tornado==6.1
traitlets==5.0.5
Twisted==20.3.0
typing-extensions==3.7.4.3
tzlocal==2.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
urwid==2.1.2
vboxapi==1.0
virtualenv==20.4.0+ds
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wfuzz==3.1.0
widgetsnbextension==2.0.0
wxPython==4.0.7
xcffib==0.8.1
xdg==5
XlsxWriter==1.1.2
yarl==1.6.3
youtube-dl==2021.2.10
zipp==1.0.0
zope.event==4.4
zope.interface==5.2.0
